+++ documentation/example/app/api_auth/locals.tf
@@ -2,7 +2,6 @@ locals {
   api_name                = "auth_api.example.com"
   auth_key_cognito        = "cognito"
   auth_key_lambda         = "lambda"
-  client_app_name         = "gen_pop"
   lambda_name_auth        = "auth"
   lambda_name_integration = "integration"
   lambda_name_options     = "options"
@@ -10,10 +9,8 @@ locals {
   output_data = {
     api    = module.api.data
     lambda = module.lambda.data
-    pool   = module.user_pool.data
     role   = module.api_role.data
   }
-  pool_name = "app"
   stage_map = {
     prd = {}
   }