+++ rds/instance/locals.tf
@@ -1,10 +1,10 @@
 module "name_map" {
   source                          = "../../name_map"
-  name_include_app_fields_default = var.db_name_include_app_fields_default
-  name_infix_default              = var.db_name_infix_default
+  name_include_app_fields_default = var.name_include_app_fields_default
+  name_infix_default              = var.name_infix_default
   name_map                        = local.l0_map
-  name_prefix_default             = var.db_name_prefix_default
-  name_suffix_default             = var.db_name_suffix_default
+  name_prefix_default             = var.name_prefix_default
+  name_suffix_default             = var.name_suffix_default
   std_map                         = var.std_map
 }