+++ secret/init_value/main.tf
@@ -1,6 +1,7 @@
 module "initial_password" {
-  source     = "../../random/password"
-  random_map = local.create_init_password_map
+  source                               = "../../random/password"
+  random_map                           = local.create_init_password_map
+  random_special_character_set_default = var.secret_random_special_character_set_default
 }
 
 module "initial_tls_key" {