+++ sm/secret/main.tf
@@ -10,12 +10,18 @@ resource "aws_secretsmanager_secret" "this_secret" {
 }
 
 module "initial_value" {
-  source     = "../../secret/init_value"
-  secret_map = local.create_init_value_map
+  source                                      = "../../secret/init_value"
+  secret_map                                  = local.create_init_value_map
+  secret_random_special_character_set_default = var.secret_random_special_character_set_default
 }
 
 resource "aws_secretsmanager_secret_version" "initial_value" {
-  for_each      = local.create_init_value_map
+  for_each = local.create_init_value_map
+  lifecycle {
+    ignore_changes = [
+      secret_string, # This is the initial value; rotate externally after
+    ]
+  }
   secret_id     = aws_secretsmanager_secret.this_secret[each.key].id
   secret_string = module.initial_value.secret_map[each.key]
 }